services/pricing-service/pricing-service/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-dotenv==1.0.0
requests==2.32.3